+++ include/TMX/TmxPolygon.h
@@ -29,7 +29,7 @@
 
 #include <vector>
 
-#include "TmxPoint.h"
+#include "TMX/TmxPoint.h"
 
 namespace tinyxml2 {
     class XMLNode;
@@ -52,7 +52,7 @@ namespace Tmx
         const Tmx::Point &GetPoint(int index) const { return points[index]; }
 
         // Get the number of vertices.
-        int GetNumPoints() const { return (int)points.size(); }
+        int GetNumPoints() const { return points.size(); }
 
     private:
         std::vector< Tmx::Point > points;
