+++ include/TMX/TmxTile.h
@@ -25,8 +25,8 @@
 
 #include <vector>
 
-#include <TMX/TmxPropertySet.h>
-#include <TMX/TmxImage.h>
+#include "TMX/TmxPropertySet.h"
+#include "TMX/TmxImage.h"
 
 namespace tinyxml2
 {
@@ -70,7 +70,7 @@ namespace Tmx
         // Returns the number of frames of the animation. If the tile is not animated, returns 0.
         int GetFrameCount() const
         {
-            return (int)frames.size();
+            return frames.size();
         }
 
         // Returns the total duration of the animation, in milliseconds,
@@ -118,7 +118,7 @@ namespace Tmx
         // Get the number of objects in the list.
         int GetNumObjects() const
         {
-            return (int)objects.size();
+            return objects.size();
         }
 
     private:
